sockets working for live chatting between users
